feat: add user allowances and group allowances querier
